Features added (HTML_TEXT): Added new notion of just "raw text" as a "TAG"
* HTML_TEXT: Represents the notion of just raw text that can be inserted
anywhere in the HTML structure. As a descendent of HTML_TAG, this class
can also have subordinate or enclosed HTML (e.g. see the test where we
inject an href anchor).
* HTML_A: Renamed back.
* HTML_A_HYPERLINK: Ibid.
* HTML_TAG: Has a place for simple raw text (e.g. `text_content'). Also,
if the `tag_name' is empty (as it will be for raw text like HTML_TEXT,
then we do not need tag-enclosing angle-brackets.
* HTML_TAG_TEST_SET: Added a test of the HTML_TEXT and immediately
encountered a bug in the Framwork FW_ATTRIBUTE_HELPER (see that lib
check-in notes).
* TEST_SETs: Renaming.