Feature added (Notes): Added more documentation on inheritance
* HTML_ATTRIBUTES: Added more documentation on how the inheritance model
is used to export the appropriate HTML_TAG attributes to Clients in
descendent classes. This allows each descendent class to export only the
attribute features that it specifically has in the HTML specification.
All other attribute features are secret (non-exported) and therefore not
available to be changed. As such, they are in the default state, which
means they will not be rendered to the HTML tag, nor will they be picked
up in the CSS-extraction visitor process, when the CSS is built from the
HTML tag. Nor will they be available to the Theming visitor to be set.